Add debug output
Otherwise, it just does what it does (not much, yet :-S), and prints
mostly nothing. (It does seem to print the channels it's trying
to join, though...) Now, if in debug mode, it logs what sub init
is doing.

Debug mode can be enabled by passing debug => 1 in the config file.