Update config file example and documentation to actual use
The example-cvnircbotrc was missing the needed nagios_logfile
and nagios_channels settings. Also ssl and debug settings are now
contained in the example. (SSL being turned off to be able to connect
to a regular server normally; debug turned on so the bot will produce
some output instead of silence.)

The last line now ends in a comma, too, which is legal but optional,
and serves to be able to easily add/remove/comment lines without
having to change unaffected lines just because they didn't end in ','
or would now do while being the new last line.

The README.md now should explain all the things the user has to take
care of (in setting the bot rc file up), in the Setup / Basic setup
section.